Numerous updates:
Revert AbstractConversationModel.getUrlProperty method. This is not the right place for this, and it is not really needed anyway.

Refactored the way that we check responses for XSS and CRLF injection. Created a couple of extra methods to simplify this.
Added checking for POST injection, IFF it is a normal (non-multipart) form.
Changed it to mark Conversations as suspicious, not URL's.
Changed the way that we identify Suspicious, and Vulnerable. Now it is separated for XSS and CRLF. While we are marking them as suspicious, we record which parameters we thought were interesting as a conversation property
Updated the way that we construct URL's to test, to pass only the necessary info.
We don't record which conversations/parameters we have tested, since it doesn't really make sense. What happens if we change the test strings, for one? Granted, this does mean that we can test the same combination of URL's over and over. Be more careful in your selection from the "suspected" table ;-)
Deleted XSSCRLFConversationTableModel. Implemented within the XSSCRLFPanel, so we can pass the columns to the SummaryPanel as well.

NOTE: In the future, we might want to create separate Test XSS and Test CRLF buttons in the UI.