Fix a reported Cross Site Scripting vulnerability.
From info@moritz-naumann.com:

ear Rogan Dawes,

this is to advise you that we have become aware of security issue(s) in
the following softwares you develop, maintain or publish:

WebScarab version 20060621-0003 and earlier versions

The following security related issues have been detected:

1. Cross Site Scripting vulnerability in error messages

By accessing the following URI using a web browser which is prone to
this issue and configured to proxy through WebScarab, a non-persitent
web script injection can be achieved:

http://x.y/</pre><script>alert(0);</script>

This allows for disclosure of sensitive data stored in the security
context of any arbitrary domain x.y which the web browser has previously
accessed but WebScarab is not able to access by the time the attack
takes place (due to invalid upstream proxy setting on WebScarab,
different results of DNS queries, limited connectivity or other reasons).

Ms Internet Explorer 6 SP2 and Konqueror 3.5.3 are known to be prone to
this issue. This problem is caused by insufficient santitation of user
supplied input before it is returned to the client as part of an error
message.

----- end excerpt

I was unable to duplicate this in brief testing, however, it makes sense
to sanitise anything we generate that goes back to the browser.