TextPanel: Be more careful about character sets.
When we get bytes to display as text, we should use the correct character encoding to convert
bytes to a string and vice versa. Until now, we have simply been choosing UTF-8 arbitrarily.

Now we use the Mozilla CharDet library to automatically detect the character set if it is not
specified on the content-type header.