Lite refactoring
Combine shared code for deciding on the Lite or standard interfaces, and
starting up the UI. Still needed is a menu option to enable or disable
the lite interface. In the meantime, it can be controlled on the command
line using:

    $ java -DWebScarab.lite=false -jar webscarab.jar

to turn off the lite version, and use the previously standard interface.

Alternatively, a properties entry WebScarab.lite=false can be set in the
WebScarab.properties file to make this persistent. We will shortly add
a menu option in the UI to set and unset this option in the props file.