Allow reading of site-specific SSL keys from the working directory
The algorithm is as follows:

When WebScarab receives a CONNECT request for www.example.com, it looks
for a file named www.example.com.p12 containing a private key and a
signed certificate. It is expected that the certificate will have a
suitable CN field so that the client will accept the cert as valid.

Note that you will need to import the cert into your client (browser,
thick client, whatever) so that it will accept the CA that signed it.

There are several utilities which are designed to provide a simple CA
facility, such as OpenSSL, or scripts that wrap OpenSSL.

An example script which *MAY* be suitable is provided in the doc/ directory.