Do not drop unrecognized Authorization headers
The Authorization may contain schemes other than Basic. Before this patch,
"Authorization: GoogleLogin ..." was silently dropped which broke an
application.

After this patch, constructAuthenticationHeader still return null when
challenges are given but negotiation fail (when headers WWW-Authenticate
or Proxy-Authenticate are given). But instead of returning null when
the authorization scheme is unrecognized, return the credentials string (the
string that is given to Proxy-Authorization or Authorization).