Add wrapper classes for InputStream and Socket
This is in preparation of supporting SNI when a client specifies an IP address
in a CONNECT request, instead of a DNS host name.

The `CONNECT ip:port HTTP/1.0` behavior was observed with certain Android
clients. In order to reproduce it locally, I setup a socat instance that
transparently proxies all requests to WebScarab with an IP address.

    socat TCP-LISTEN:443,fork PROXY:127.0.0.1:178.21.112.251:443,proxyport=8008
    curl -4 -vvv -k --resolve lekensteyn.nl:443:127.0.0.1 https://lekensteyn.nl