Enable SNI support for WebScarab in the server role
Before this patch, WebScarab will always create a certificate based on the host
given in the CONNECT line. Some clients send an IP address in the CONNECT line
and put the domain name in the TLS ClientHello message. This causes a mismatch
between the name on the certificate (IP address) and the expected value (a
domain name).

Since the host name passed to the negiotiateSSL (SSLSocketFactory.createSocket)
is used for the certificate name, the real host name must be determined before
that (before the SSLSocket is set up). Therefore, wrap the socket in a
WireSocket such that the ClientHello message can be captured and the Server Name
can be extracted.

Once a SNI host name is found, the `_base` URL is rebased on this host name.
This changes the certificate name and replaces the host that is visible in the
UI and logs.